+++ CHANGELOG.md
@@ -1,8 +1,13 @@
 ## 0.0.1
 
-* TODO: Describe initial release.
+* Integrate app state for android
 
 ## 1.0.0
 
 * Update README.md
 * Integrate app state for iOS
+
+## 1.0.1
+
+* Update README.md
+* fixed incorrect return type of state